refactor mobile-client utils: add MOBILE_DISPLAY_NAME_PATTERN for display name validation and update exports in index file for better organization
