+++ packages/mobile-client/src/utils/index.ts
@@ -1,2 +1,2 @@
+export * from './patterns';
 export * from './serverResponseGuards';
-export * from './validations';
