+++ packages/mobile-client/src/utils/patterns.ts
